fix: u-boot: Add support for non-secure efuse for Rock64
This change uses a series of patches from u-boot that still haven't made
their way into the main repository.

Essentially what it does, is that it adds support for the non-secure
efuse, which allows u-boot to generate a stable mac address derived
from the cpu id.

All credit goes to the original patch authors.

https://patchwork.ozlabs.org/project/uboot/patch/20201013202116.24850-2-jonas@kwiboo.se/
https://patchwork.ozlabs.org/project/uboot/patch/20201013202116.24850-3-jonas@kwiboo.se/
https://patchwork.ozlabs.org/project/uboot/patch/20201013202116.24850-4-jonas@kwiboo.se/
Signed-off-by: Jan Praczyk <jpraczyk@users.noreply.github.com>